server/python_bot/requirements.txt
flask==3.0.0
flask_cors==4.0.0
python-dotenv==1.0.0
openai>=1.0.0
sentence-transformers>=2.0.0
numpy>=1.24.0
